Allow tests to be marked "doctest: +XFAIL"
At least for now, this merely inverts the pass logic.  Which is
useful to implement tests that are supposed to fail (for which its
a failure if they pass).  Note however, that this behaviour is
slightly different from the Octave `test` XFAIL behaviour so is
possibly subject to change in the future.

Fixes #67