Fix for Matlab/Octave regex difference
I'm not sure which is "correct" but Octave will give a token for a
`()` group even if placed inside a `(?:)` group.  Matlab will not.
Fix is to just grab the white space and the parentheses in the token
with a single group.  This works because I think its harmless for the
code to be wrapped in parentheses (and extra whitespace).  Fixes #162.