workaround arm-specific regexprep bug
This fixes #130.  Again.  To workaround, we avoid using regexprep and
instead get the tokens with regexp and do the substitution ourselves.
Leave the regexprep code commented out with a link to the newly-submitted
upstream bug.  This can be undone once a fix lands upstream.
Prep for a new release with a minor version bump.