[maint] run CI tests on distro Octave, PPA and daily snapshots
Support testing on multiple Octave versions.  Tested with xenial
but snapshots are not (yet) available there.  Left some notes to
future-proof this a bit.

It looks like the octave package install can be factored out but
I saw this warning in the Travis docs:

> Note: When using APT sources and packages together, you
> need to make sure they are under the same key space in
> the YAML file.

https://docs.travis-ci.com/user/installing-dependencies/

Probably distro will fail b/c it will be ancient, but again its
future-proofing a bit.

Snapshot needs some deps the distro: easiest to just install octave,
and it seems the PPA is not needed for this.