track errors until compare optionally prefix with "error:"
If we label expected errors as such within our data structure then
we can be more flexible at compare time.  For example, if the user
wants "error: " we can make that optional for matching.  This is
helpful b/c Octave puts "error: " at the beginning of some error
messages.  Matlab typically does not and yet this allows Matlab
docs to also start with "error: ".

We also support starting with "??? ", but this is deprecated and
not recommended.  Fixes #238.